fixed install script

+++ install.sh
@@ -1,6 +1,8 @@
 #!/bin/bash
 systemctl disable management.service
-rm /etc/systemd/system/management.service
+if ! -e /etc/systemd/system/management.service; then
+    rm /etc/systemd/system/management.service
+fi
 mv ./management.service /etc/systemd/system
 systemctl start management.service
 systemctl enable management.service
